Make options name function public
